Bloc provider added
