+++ pubspec.yaml
@@ -29,6 +29,7 @@ dependencies:
     equatable: ^2.0.0
     http: ^0.13.1
     formz: ^0.3.2
+    flutter_dotenv: ^3.1.0
 
 dev_dependencies:
     flutter_test:
@@ -45,9 +46,8 @@ flutter:
     uses-material-design: true
 
     # To add assets to your application, add an assets section, like this:
-    # assets:
-    #   - images/a_dot_burr.jpeg
-    #   - images/a_dot_ham.jpeg
+    assets:
+        - .env
 
     # An image asset can refer to one or more resolution-specific "variants", see
     # https://flutter.dev/assets-and-images/#resolution-aware.
